backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
transformers
torch
sentencepiece
scikit-learn
beautifulsoup4
requests
pydantic-settings
pydantic[email]
python-multipart

# Database & Autentikasi
motor
pymongo[srv]
passlib==5.3.0
bcrypt==3.2.2
python-jose[cryptography]